requirements.txt
gunicorn==21.2.0
flask==2.1.1
oauth2client==4.1.3
gspread==4.0.1
premailer==3.10.0
Pillow==8.2.0
pip>=23.2.1
pycryptodome==3.11.0